+++ Action/RangeAttackPlayer.cpp
@@ -14,7 +14,7 @@ RangeAttackPlayer::~RangeAttackPlayer()
 {
 }
 
-void RangeAttackPlayer::Attack(const Vector3 & _ownerPosition,const int& _direction)
+void RangeAttackPlayer::Attack(const Vector3 & _ownerPosition,const float& _direction)
 {
 	new ThrowWeapon(_ownerPosition, _direction,20);
 }
